+++ modules/cactus-web/src/MenuBar/MenuBar.story.tsx
@@ -1,8 +1,10 @@
-import { number } from '@storybook/addon-knobs'
+import { number, select } from '@storybook/addon-knobs'
 import { Meta } from '@storybook/react/types-6-0'
 import React from 'react'
 
-import MenuBar from './MenuBar'
+import MenuBar, { MenuBarVariants } from './MenuBar'
+
+const menuBarVariants: MenuBarVariants[] = ['light', 'dark']
 
 const LABELS = [
   'Ready!',
@@ -67,6 +69,43 @@ export const BasicUsage = (): React.ReactElement => {
     }
     return (
       <Component key={ix} {...props}>
+        {items}
+      </Component>
+    )
+  }
+
+  return <>{makeList(totalDepth, {}, 0, MenuBar)}</>
+}
+
+export const MenuBarDark = (): React.ReactElement => {
+  const breadth = number('Breadth', 8)
+  const totalDepth = number('Depth', 2)
+  const variantSelection = select('variant', menuBarVariants, 'dark')
+
+  const makeList = (
+    depth: number,
+    props: any,
+    ix: number,
+    Component: React.ComponentType<any> = MenuBar.List
+  ) => {
+    const items = []
+    for (let i = 0; i < breadth; i++) {
+      // We need the default to be 100% reproducible for storyshots.
+      const useStatic = depth === totalDepth && i < 8
+      const useList = useStatic ? i % 3 === 0 : (Math.random() * depth) / 1.5 > 0.5
+      const title = useStatic ? LABELS[i] : getLabel()
+      if (useList) {
+        items.push(makeList(depth - 1, { title }, i))
+      } else {
+        items.push(
+          <MenuBar.Item key={i} onClick={action(`x: ${i}, y: ${totalDepth - depth}`)}>
+            {title}
+          </MenuBar.Item>
+        )
+      }
+    }
+    return (
+      <Component key={ix} variant={variantSelection} {...props}>
         {items}
       </Component>
     )