+++ modules/cactus-web/src/Modal/Modal.tsx
@@ -5,6 +5,7 @@ import React, { FunctionComponent } from 'react'
 import styled, { css } from 'styled-components'
 import { height, HeightProps, maxHeight, MaxHeightProps, width, WidthProps } from 'styled-system'
 
+import { DimmerStyled } from '../Dimmer/Dimmer'
 import Flex from '../Flex/Flex'
 import { border, boxShadow, radius } from '../helpers/theme'
 import cssVariant from '../helpers/variant'
@@ -48,10 +49,11 @@ const Modalbase: FunctionComponent<ModalProps> = (props): React.ReactElement => 
       variant={variant}
       height={innerHeight}
       maxHeight={innerMaxHeight}
+      as={DialogOverlay}
       {...rest}
     >
       <DialogContent aria-label={modalLabel}>
-        <IconButton onClick={onClose} label={closeLabel}>
+        <IconButton className="modal-close-btn" onClick={onClose} label={closeLabel}>
           <NavigationClose />
         </IconButton>
         {hasChildren && (
@@ -69,13 +71,12 @@ const Modalbase: FunctionComponent<ModalProps> = (props): React.ReactElement => 
   )
 }
 
-export const ModalPopUp = styled(DialogOverlay).withConfig({
+export const ModalPopUp = styled(DimmerStyled).withConfig({
   shouldForwardProp: (prop) => {
     // @ts-ignore
     return prop !== 'maxHeight'
   },
 })<ModalPopupProps>`
-  background: rgba(46, 53, 56, 0.33);
   bottom: 0;
   display: flex;
   left: 0;
@@ -84,6 +85,7 @@ export const ModalPopUp = styled(DialogOverlay).withConfig({
   right: 0;
   top: 0;
   align-items: center;
+  flex-direction: row;
   z-index: 101;
   > [data-reach-dialog-content] {
     ${(p) => p.width && 'box-sizing: border-box;'}
@@ -93,7 +95,6 @@ export const ModalPopUp = styled(DialogOverlay).withConfig({
     border-radius: ${radius(20)};
     background: white;
     ${(p): string => boxShadow(p.theme, 2)};
-    margin: auto;
     max-width: ${(p) => !p.width && '80%'};
     ${width}
     outline: none;
@@ -125,7 +126,7 @@ export const ModalPopUp = styled(DialogOverlay).withConfig({
         border-color: ${(p): string => p.theme.colors.error};
       `,
     })}
-    ${IconButton} {
+    .modal-close-btn {
       height: 16px;
       position: absolute;
       right: 24px;
@@ -136,7 +137,7 @@ export const ModalPopUp = styled(DialogOverlay).withConfig({
     ${(p): string | undefined => p.theme.mediaQueries && p.theme.mediaQueries.medium} {
       padding: 40px 88px;
       max-width: ${(p) => !p.width && '30%'};
-      ${IconButton} {
+      .modal-close-btn {
         height: 24px;
         position: absolute;
         right: 40px;